migrate simapp tests to use main app - WIP
