+++ app/genesis.go
@@ -2,9 +2,67 @@ package app
 
 import (
 	"encoding/json"
+	"io"
+	"os"
 
 	"github.com/cosmos/cosmos-sdk/codec"
+	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
+	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	tmjson "github.com/tendermint/tendermint/libs/json"
+	tmtypes "github.com/tendermint/tendermint/types"
 )
+
+// AppStateFromGenesisFileFn util function to generate the genesis AppState
+// from a genesis.json file.
+func AppStateFromGenesisFileFn(r io.Reader, cdc codec.JSONCodec, genesisFile string) (tmtypes.GenesisDoc, []simtypes.Account) {
+	bytes, err := os.ReadFile(genesisFile)
+	if err != nil {
+		panic(err)
+	}
+
+	var genesis tmtypes.GenesisDoc
+	// NOTE: Tendermint uses a custom JSON decoder for GenesisDoc
+	err = tmjson.Unmarshal(bytes, &genesis)
+	if err != nil {
+		panic(err)
+	}
+
+	var appState GenesisState
+	err = json.Unmarshal(genesis.AppState, &appState)
+	if err != nil {
+		panic(err)
+	}
+
+	var authGenesis authtypes.GenesisState
+	if appState[authtypes.ModuleName] != nil {
+		cdc.MustUnmarshalJSON(appState[authtypes.ModuleName], &authGenesis)
+	}
+
+	newAccs := make([]simtypes.Account, len(authGenesis.Accounts))
+	for i, acc := range authGenesis.Accounts {
+		// Pick a random private key, since we don't know the actual key
+		// This should be fine as it's only used for mock Tendermint validators
+		// and these keys are never actually used to sign by mock Tendermint.
+		privkeySeed := make([]byte, 15)
+		if _, err := r.Read(privkeySeed); err != nil {
+			panic(err)
+		}
+
+		privKey := secp256k1.GenPrivKeyFromSecret(privkeySeed)
+
+		a, ok := acc.GetCachedValue().(authtypes.AccountI)
+		if !ok {
+			panic("expected account")
+		}
+
+		// create simulator accounts
+		simAcc := simtypes.Account{PrivKey: privKey, PubKey: privKey.PubKey(), Address: a.GetAddress()}
+		newAccs[i] = simAcc
+	}
+
+	return genesis, newAccs
+}
 
 // GenesisState of the blockchain is represented here as a map of raw json
 // messages key'd by a identifier string.
