+++ x/testutil/cli/network.go
@@ -127,7 +127,7 @@ type (
 )
 
 // BuildNetworkConfig returns a configuration for a local in-testing network
-func BuildNetworkConfig(appGenesis simapp.GenesisState) Config {
+func BuildNetworkConfig(appGenesis app.GenesisState) Config {
 	encCfg := app.MakeTestEncodingConfig()
 
 	return Config{
